Moved ClientMethod to a separate file.
